+++ example/lib/main.dart
@@ -70,14 +70,27 @@ class HomePage extends StatelessWidget {
               onTap: () {
                 AVSImageGallery(
                   context,
+                  closeButtonPosition: CloseButtonPosition.topLeft,
+                  customCloseButton: const Icon(Icons.delete, color: Colors.red),
                   imagePaths: [
                     "assets/image1.png",
-                    "assets/image2.png",
+                    "https://nightgoes.com/wp-content/uploads/2023/09/pro.gaultier_a_modern_electric_night_train_crossing_the_map_of__8751c62e-5aa0-4af7-8138-dddf21d8141b.png.webp",
+                    "https://nightgoes.com/wp-content/uploads/2023/09/paris-berlin-1.png.webp"
+                        "assets/image2.png",
                     "assets/image3.png",
                     "assets/image4.png",
                     "assets/image5.png",
                   ],
-                  initialIndex: 2,
+                  imageGalleryStyle: ImageGalleryStyle(
+                    slideHeight: 18,
+                    slideWidth: 18,
+                    activeSlideColor: Colors.red,
+                    inActiveSlideColor: Colors.amber,
+                    activeSlideIcon: "assets/star-solid.svg",
+                    inActiveSlideIcon: "assets/star-regular.svg",
+                  ),
+                  
+                  // initialIndex: 2,
                   backgroundColor: Colors.red,
                   backgroundGradient: const LinearGradient(
                     colors: [Colors.blue, Color.fromARGB(255, 176, 200, 91)],
@@ -108,9 +121,7 @@ class HomePage extends StatelessWidget {
                 color: Colors.grey[300],
                 image: DecorationImage(
                   fit: BoxFit.contain,
-                  image: AVSImageProvider(
-                      "https://www.svgrepo.com/show/530440/machine-vision.svg",
-                      scale: 9),
+                  image: AVSImageProvider("https://www.svgrepo.com/show/530440/machine-vision.svg", scale: 9),
                 ),
               ),
               child: const Text("SVG Image Provider"),