+++ lib/functions/_check_local.dart
@@ -3,7 +3,9 @@ part of "../avs_image.dart";
 AVSImageType _getImageType(String path) {
   if (path.startsWith('http://') || path.startsWith('https://')) {
     return AVSImageType.network;
-  } else if (path.startsWith('/') || path.contains('/storage/') || path.contains('/data/')) {
+  } else if (path.startsWith('/') ||
+      path.contains('/storage/') ||
+      path.contains('/data/')) {
     return AVSImageType.file;
   } else {
     return AVSImageType.asset;